// src/api/generated/models/getCompaniesCompanyIdShippingInvoicesDirection.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type GetCompaniesCompanyIdShippingInvoicesDirection = typeof GetCompaniesCompanyIdShippingInvoicesDirection[keyof typeof GetCompaniesCompanyIdShippingInvoicesDirection];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesCompanyIdShippingInvoicesDirection = {
  inbound: 'inbound',
  outbound: 'outbound',
} as const;