// src/api/generated/models/shippingInvoiceSummaryStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type ShippingInvoiceSummaryStatus = typeof ShippingInvoiceSummaryStatus[keyof typeof ShippingInvoiceSummaryStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ShippingInvoiceSummaryStatus = {
  DRAFT: 'DRAFT',
  ISSUED: 'ISSUED',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
} as const;